+++ oB-WAF.py
@@ -4,6 +4,8 @@
 Written by oBreak.
 
 Disclaimer: This tool should not be referenced in legal proceedings.
+
+Source code should be found at https://github.com/oBreak/oB-WAF-verifier
 '''
 
 # Imports
@@ -47,18 +49,24 @@ debugfileWin    = os.path.join(fileDir, 'debug\\debug.txt')
 mypathWin       = home + "\\Projects\\gh-oB-WAF-Verifier\\in\\"
 outfileWin      = os.path.join(fileDir, 'out\\out.txt')
 
-targets         = []
+sites         = []
 fileCount       = int(0)
 conf            = {}
 lineBreaks      = str('\n')
 scanresults     = []
-totalTargetsAndScans = int(0)
+totalsitesAndScans = int(0)
 percentageComplete = int(0)
 
 
 
 
 # Functions
+
+'''
+inbound() is intended to review all files from the relative /in/ directory that are of type ".txt", ".log", or ".csv".
+These files are added to the inbound file list (variable iF). If files are outside of specified file types, it
+does not add them to the inbound file list. Inbound file list is then returned by the function.
+'''
 
 def inbound():
     global fileCount, mypath
@@ -91,44 +99,51 @@ def inbound():
             fileCount = fileCount + 1
         mypath = mypathMac
     if debugSet == True:
-        debugLog.append('iF (inbound files) contains:')
+        debug.append('\t\tiF (inbound files) contains:')
         for x in iF:
-            debugLog.append(x + ' added to iF')
-        debugLog.append('fileCount is: ' + str(fileCount))
+            debug.append('\t\t' + x + ' added to inbound files list variable (iF)')
+        debug.append('\t\tfileCount is: ' + str(fileCount))
     # for t in iF:
     #     print(t, "in inbound()")
     return iF
 
 
-def parse(x):
-    file = open('in/' + x, 'r')
+'''
+parse() is intended to pull the websites from files in the inbound file list (iF), which is created in inbound(), 
+and adds them to the sites list. The sites list is a global variable that is used by the scrape() function.
+'''
+def parse(iF):
+    file = open('in/' + iF, 'r')
     for line in file:
         line = line.strip()  # Removes blank lines.
         if line:
             ''' This portion of the script is meant to discover if the inbound files are correctly formatted.
             Decision tree:
-                If IP -> Add IP to target list (in other languages, this is an array)
+                If IP -> Add IP to sites list (in other languages, this is an array)
                 If CIDR -> Convert to IPs
-                    -> Add all IPs to targets list
+                    -> Add all IPs to sites list
                 If IP Range -> Convert to IPs
-                    -> Add all IPs to targets list
+                    -> Add all IPs to sites list
 
 
             if line == IP address format:
-                targets.append(line)
+                sites.append(line)
                 if debugSet == True:
-                debugLog.append('Appending ' + line + ' to targets.')
+                debugLog.append('Appending ' + line + ' to sites.')
             '''
-            targets.append(line)
+            sites.append(line)
             if debugSet == True:
-                debugLog.append('Appending ' + line + ' to targets.')
+                debugLog.append('Appending ' + line + ' to sites.')
     pass
 
-
-def scrape(iF):
+'''
+scrape() is intended to get the response.text from each of the sites identified in the parse() function.
+The response text is then added to the out list, which is used by the webout() function.
+'''
+def scrape():
     global out
     #urls = ['http://www.google.com', 'https://www.espn.com']
-    for t in targets:
+    for t in sites:
         # print(t)
         try:
             response = requests.get(t)
@@ -141,6 +156,10 @@ def scrape(iF):
     debug.append('\tscrape() end')
     return
 
+'''
+webout() is intended to print the response.text from each of the sites identified to an out file in the 
+relative /out/ folder.
+'''
 
 def webout():
     # Set local variables
@@ -176,6 +195,12 @@ def webout():
     print('\tOutput complete: ' + str(n) + ' lines written.\n')
     return
 
+'''
+debugout() is intended to print the debug log for the program.
+The debug list variable contains all log entries.
+This might be a candidate for concurrent processing, as it won't complete if the program breaks midway.
+'''
+
 def debugout():
     # Set local variables
     lineBreaks = '\n'        # This is making sure the outfile has separate lines.
@@ -210,6 +235,16 @@ def debugout():
     print('\tOutput complete: ' + str(n) + ' lines written.\n')
     return
 
+'''
+main() is the main program. Workflow is:
+
+- Get list of files with sites to review
+- Extract sites from list of files
+- For each site in sites, requests -> site; append to out list
+- Print debug log to debug file.
+- Print out list to out file.
+- End.
+'''
 
 def main():
     debug.append('main() start')
@@ -223,15 +258,15 @@ def main():
         parse(t)
     debug.append('\tparse() end')
     debug.append('\tscrape() start')
-    for t in targets:
-        scrape(t)
+    for site in sites:
+        scrape()
     debug.append('\tscrape() end')
     debug.append('main() end')
-    debugout()
+    if debugSet == True:
+        debugout()
     webout()
     return
 
+# Running the program
+
 main()
-
-
-# debug()